Cleaned up (and fixed previously broken changes) for ChestCloseEvent

+++ src/main/java/co/neweden/LootCrates/listeners/PlayerListener.java
@@ -39,49 +39,45 @@ public class PlayerListener implements Listener {
         Chest c = (Chest) event.getInventory().getHolder();
 
         //check if chest exists
-        Database.ChestClass chest = Database.getCrateFromHashMap(c.getBlock()); //Checks if chest is found
+        Database.ChestClass chest = Database.getCrateFromHashMap(c.getBlock());
         if (chest == null) return;
 
+        // We need to know if the chest is actually empty or not
+        int itemsLeft = 0;
+        for (ItemStack item : event.getInventory().getContents()) {
+            if (item != null) itemsLeft++;
+        }
+
+        if (itemsLeft > 0 && !chest.found) {
+            // Chest was found but still has items
+            String FoundChest_NB_Colored = ChatColor.translateAlternateColorCodes('&', ConfigRetriever.FoundChest_NB);
+            player.sendMessage(FoundChest_NB_Colored);
+            Timer.OnCrateCreated(c.getBlock(), 6000); //6000=5min, 600=30sec
+        }
+
+        if (itemsLeft <= 0) {
+            // All items were removed from the chest
+            if (!chest.found) { // Chest was found
+                String FoundChest_Colored = ChatColor.translateAlternateColorCodes('&', ConfigRetriever.FoundChest);
+                player.sendMessage(FoundChest_Colored);
+            }
+            Database.removeChest(c.getBlock());
+            c.getLocation().getBlock().setType(Material.AIR);
+            ChestSpawner.newChest();
+        }
+
+        if (chest.found) return;
+
+        // Construct broadcast message
         String specialText = "";
         if (chest.tier == 1) {
             specialText = ChatColor.WHITE + " | " + ChatColor.RED + "(╯°□°）╯︵ ┻━┻";
         }
         String message = ChatColor.GOLD + player.getDisplayName() + ChatColor.GRAY + " has found a " + ChatColor.YELLOW + ChestSpawner.tierCalc(chest.tier) + ChatColor.GRAY + " Crate" + specialText;// + " in " + player.getWorld().getName();
+        Bukkit.broadcastMessage(message);
 
-//        if (event.getInventory().getContents().length > 0 && chest.found) {
-//            //Chest was found by a player | marking it as found
-//            Database.chestIsFound(player,chest);
-//            Bukkit.broadcastMessage(message);
-//            String FoundChest_NB_Colored = ChatColor.translateAlternateColorCodes('&', ConfigRetriever.FoundChest_NB);
-//            player.sendMessage(FoundChest_NB_Colored);
-//            Timer.OnCrateCreated(c.getBlock(), 6000); //6000=5min, 600=30sec
-//            return;
-//        } This is broken
-
-        ItemStack[] items = event.getInventory().getContents();
-        for (ItemStack item : items) {
-            if (item == null) {continue;}
-            //Checks if chest has already been found so it wont display a message twice
-            if (chest.found) {return;}
-            //Chest was found by a player | marking it as found
-            Database.chestIsFound(player,chest);
-            Bukkit.broadcastMessage(message);
-            String FoundChest_NB_Colored = ChatColor.translateAlternateColorCodes('&', ConfigRetriever.FoundChest_NB);
-            player.sendMessage(FoundChest_NB_Colored);
-            Timer.OnCrateCreated(c.getBlock(), 6000); //6000=5min, 600=30sec
-            return;
-        }
-
-        //Run code to execute if the chest is empty
-        if(!chest.found){
-            Database.chestIsFound(player,chest);
-            Bukkit.broadcastMessage(message);
-            String FoundChest_Colored = ChatColor.translateAlternateColorCodes('&', ConfigRetriever.FoundChest);
-            player.sendMessage(FoundChest_Colored);
-        }
-        Database.removeChest(c.getBlock());
-        c.getLocation().getBlock().setType(Material.AIR);
-        ChestSpawner.newChest();
+        // mark chest as found and update player statistics
+        Database.chestIsFound(player,chest);
     }
 
 }
